Move more from sound to PlaybackContainer
