+++ src/sound.ts
@@ -20,12 +20,14 @@
  * and can be controlled individually. This architecture allows for complex audio behaviors, such as playing multiple overlapping
  * instances of a sound with different settings, without requiring the user to manually manage each playback instance.
  */
+
+
 import { PlaybackContainer } from "container";
+import { AudioContext, IAudioBuffer, IPannerOptions } from "standardized-audio-context";
 import { BaseSound, LoopCount, PanType, Position, SoundType } from "./cacophony";
 import { BiquadFilterNode, GainNode, SourceNode, } from './context';
 import { FilterManager } from "./filters";
 import { Playback } from "./playback";
-import { AudioContext, IAudioBuffer, IPannerOptions } from "standardized-audio-context";
 
 
 type SoundCloneOverrides = {
@@ -43,38 +45,14 @@ type SoundCloneOverrides = {
 export class Sound extends PlaybackContainer(FilterManager) implements BaseSound {
     buffer?: IAudioBuffer;
     context: AudioContext;
-    private globalGainNode: GainNode;
-    private _position: Position = [0, 0, 0];
-    private _stereoPan: number = 0;
-    private _threeDOptions: IPannerOptions = {
-        coneInnerAngle: 360,
-        coneOuterAngle: 360,
-        coneOuterGain: 0,
-        distanceModel: 'inverse',
-        maxDistance: 10000,
-        channelCount: 2,
-        channelCountMode: 'clamped-max',
-        channelInterpretation: 'speakers',
-        panningModel: 'HRTF',
-        refDistance: 1,
-        rolloffFactor: 1,
-        positionX: 0,
-        positionY: 0,
-        positionZ: 0,
-        orientationX: 0,
-        orientationY: 0,
-        orientationZ: 0
-    };
     loopCount: LoopCount = 0;
     private _playbackRate: number = 1;
-    private _volume: number = 1;
 
-    constructor(public url: string, buffer: AudioBuffer | undefined, context: AudioContext, globalGainNode: GainNode, public type: SoundType = SoundType.Buffer, public panType: PanType = 'HRTF'
+    constructor(public url: string, buffer: AudioBuffer | undefined, context: AudioContext, private globalGainNode: GainNode, public type: SoundType = SoundType.Buffer, public panType: PanType = 'HRTF'
     ) {
         super();
         this.buffer = buffer;
         this.context = context;
-        this.globalGainNode = globalGainNode;
     }
 
     /**
@@ -170,49 +148,6 @@ export class Sound extends PlaybackContainer(FilterManager) implements BaseSound
     }
 
     /**
-    * Retrieves the current 3D spatial position of the sound in the audio context.
-    * The position is returned as an array of three values[x, y, z].
-    * @returns { Position } The current position of the sound.
-    */
-
-    get position(): Position {
-        return [this._threeDOptions.positionX, this._threeDOptions.positionY, this._threeDOptions.positionZ]
-    }
-
-    /**
-    * Sets the 3D spatial position of the sound in the audio context.
-    * The position is an array of three values[x, y, z].
-    * This method updates the position of all active playbacks of the sound.
-    * @param { Position } position - The new position of the sound.
-    */
-
-    set position(position: Position) {
-        this._threeDOptions.positionX = position[0];
-        this._threeDOptions.positionY = position[1];
-        this._threeDOptions.positionZ = position[2];
-        this.playbacks.forEach(p => p.position = position);
-    }
-
-
-    get threeDOptions(): IPannerOptions {
-        return this._threeDOptions;
-    }
-
-    set threeDOptions(options: Partial<IPannerOptions>) {
-        this._threeDOptions = { ...this._threeDOptions, ...options };
-        this.playbacks.forEach(p => p.threeDOptions = this._threeDOptions);
-    }
-
-    get stereoPan(): number | null {
-        return this._stereoPan;
-    }
-
-    set stereoPan(value: number) {
-        this._stereoPan = value;
-        this.playbacks.forEach(p => p.stereoPan = value);
-    }
-
-    /**
     * Sets or retrieves the loop behavior for the sound.
     * If loopCount is provided, the sound will loop the specified number of times.
     * If loopCount is 'infinite', the sound will loop indefinitely until stopped.
@@ -228,29 +163,6 @@ export class Sound extends PlaybackContainer(FilterManager) implements BaseSound
         this.loopCount = loopCount;
         this.playbacks.forEach(p => p.loop(loopCount));
         return this.loopCount;
-    }
-
-    /*** 
-            * Gets the volume of the sound. This volume level affects all current and future playbacks of this sound instance.
-            * The volume is specified as a linear value between 0 (silent) and 1 (full volume).
-            * 
-            * @returns {number} The current volume of the sound.
-            */
-
-    get volume(): number {
-        return this._volume;
-    }
-
-    /***
-    * Sets the volume of the sound. This volume level affects all current and future playbacks of this sound instance.
-    * The volume is specified as a linear value between 0 (silent) and 1 (full volume).
-    * 
-    * @param {number} volume - The new volume level for the sound.
-    */
-
-    set volume(volume: number) {
-        this._volume = volume;
-        this.playbacks.forEach(p => p.volume = volume);
     }
 
     get playbackRate(): number {
